#3 - 클래스명 수정 - 변수명 수
