+++ Promiss_iOS/Base.lproj/Main.storyboard
@@ -13,7 +13,7 @@
         <!--Login View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="login" id="BYZ-38-t0r" customClass="login_ViewController" customModule="Promiss_iOS" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="login" id="BYZ-38-t0r" customClass="LoginViewController" customModule="Promiss_iOS" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -62,7 +62,7 @@
                                     <color key="titleColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </state>
                                 <connections>
-                                    <action selector="clickLoginButton:" destination="BYZ-38-t0r" eventType="touchUpInside" id="aUz-qv-uZG"/>
+                                    <action selector="clickLoginButton:" destination="BYZ-38-t0r" eventType="touchUpInside" id="uav-Lq-rHA"/>
                                 </connections>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="gFR-9g-WBV">
@@ -72,17 +72,13 @@
                                     <color key="titleColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </state>
                                 <connections>
-                                    <action selector="clickSignUpButton:" destination="BYZ-38-t0r" eventType="touchUpInside" id="BZE-Jc-Bmu"/>
+                                    <action selector="clickSignUpButton:" destination="BYZ-38-t0r" eventType="touchUpInside" id="jjN-ww-5ML"/>
                                 </connections>
                             </button>
                         </subviews>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
-                    <connections>
-                        <outlet property="idTextField" destination="Vyn-nT-9L3" id="DLr-Tp-asQ"/>
-                        <outlet property="pwTextField" destination="DZ7-9B-1zK" id="hHy-Zw-kXJ"/>
-                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
             </objects>
@@ -91,7 +87,7 @@
         <!--Signup View Controller-->
         <scene sceneID="P1P-WW-Ldf">
             <objects>
-                <viewController storyboardIdentifier="signup" id="3QK-Oo-kgw" customClass="signup_ViewController" customModule="Promiss_iOS" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="signup" id="3QK-Oo-kgw" customClass="SignupViewController" customModule="Promiss_iOS" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="hMs-YY-8y9">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -154,18 +150,13 @@
                                     <color key="titleColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </state>
                                 <connections>
-                                    <action selector="clickSignUpButton:" destination="3QK-Oo-kgw" eventType="touchUpInside" id="qfu-Rd-44w"/>
+                                    <action selector="clickSignUpButton:" destination="3QK-Oo-kgw" eventType="touchUpInside" id="cgw-iA-Bkf"/>
                                 </connections>
                             </button>
                         </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                         <viewLayoutGuide key="safeArea" id="TMa-K4-kIR"/>
                     </view>
-                    <connections>
-                        <outlet property="idTextField" destination="569-Ib-Pea" id="nKt-ks-3nT"/>
-                        <outlet property="pwCheckTextField" destination="T3u-Ky-6gO" id="fHw-d4-l2f"/>
-                        <outlet property="pwTextField" destination="8BW-1X-A9g" id="ZD5-M4-by7"/>
-                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Hfe-Qo-Lna" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
