+++ Promiss_iOS/SignupViewController.swift
@@ -1,0 +1,21 @@
+//
+//  signup_ViewController.swift
+//  Promiss_iOS
+//
+//  Created by Anna Lee on 02/09/2019.
+//  Copyright © 2019 Anna Lee. All rights reserved.
+//
+
+import UIKit
+
+class SignupViewController: UIViewController {
+ 
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    @IBAction func clickSignUpButton(_ sender: Any) {
+        self.dismiss(animated: true)
+    }
+}
